*navbar , +spinner, changes to footer
