+++ tracknow/web/src/components/Loading/LoadingSpinner.tsx
@@ -1,6 +1,9 @@
 import * as React from "react";
+import { Spinner } from "@chakra-ui/react";
 
 export const LoadingSpinner = () => (
     <>
+        <Spinner color="red" />
+
     </>
 );
