+++ tracknow/web/src/components/Welcome/Welcome.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import { NavbarWelcome } from "../Navbar/NavbarWelcome";
+import { NavbarWelcome } from "../Navbar/Navbar";
 import { Hero } from "./Hero";
 import Footer from "../Footer/Footer";
